requirements.txt
python=3.10.9
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2

# Face recognition stack
click==8.2.1
colorama==0.4.6
dlib @ file:///C:/Users/learn%20%26%20Practice/Downloads/dlib-19.24.6-cp310-cp310-win_amd64.whl
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
